feat(utils): fixing linting problem wit config loader
